Replaced old MediaManager with new one
